--- client/src/components/ImageUpload/index.ts
@@ -1,3 +1,0 @@
-import ImageUpload from './ImageUpload';
-
-export default ImageUpload;
